refactor(header/footer): Extract common components and hooks
- Extract renderMenuItems and renderSocialMediaLink
- Extract useIsMobile, useOutsideClick, useScrollProgress hooks from Header
